restructure code from processing.go
